+++ src/components/Dashboard/MyOrder/MyOrder.js
@@ -3,30 +3,45 @@ import { useEffect, useState } from "react";
 import { Container } from "react-bootstrap";
 import toast from "react-hot-toast";
 import { Link } from "react-router-dom";
+import swal from "sweetalert";
 import useAuth from "../../../hook/useAuth";
 
 const MyOrder = () => {
-  const [cars, setCars] = useState([]);
+  const [cars, setCars] = useState(false);
   const { firebaseContext } = useAuth();
   const { user } = firebaseContext;
 
   useEffect(() => {
-    const url = `http://localhost:5000/orders/${user.email}`;
+    const url = `${process.env.REACT_APP_REST_API}orders/${user.email}`;
     axios.get(url).then((res) => {
       setCars(res.data);
     });
   }, [user.email]);
 
   const handleOrderDelete = (id) => {
-    const url = `http://localhost:5000/orders/${id}`;
-    axios.delete(url).then((res) => {
-      const data = res.data;
-      const filterCar = cars.filter((car) => car._id !== id);
-      if (data.acknowledged) {
-        setCars(filterCar);
-        toast.success("Successfully Delete Car!");
+    const url = `${process.env.REACT_APP_REST_API}orders/${id}`;
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this imaginary file!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        axios.delete(url).then((res) => {
+          const data = res.data;
+          const filterCar = cars.filter((car) => car._id !== id);
+          if (data.acknowledged) {
+            setCars(filterCar);
+            toast.success("Successfully Delete Car!");
+          } else {
+            toast.error("Something went wrong!");
+          }
+        });
       } else {
-        toast.error("Something went wrong!");
+        toast("Delete Cancel!", {
+          icon: <i class="fas fa-info-circle"></i>,
+        });
       }
     });
   };
@@ -34,7 +49,7 @@ const MyOrder = () => {
   return (
     <Container>
       <h2 className="text-center my-4">My Orders</h2>
-      {cars.length === 0 ? (
+      {!cars ? (
         <div
           style={{
             display: "flex",
@@ -72,35 +87,48 @@ const MyOrder = () => {
         </div>
       ) : (
         <div className="table_wrapper my-4">
-          <table>
-            <thead>
-              <tr>
-                <th>Car</th>
-                <th>Amount</th>
-                <th>Details</th>
-                <th>Status</th>
-              </tr>
-            </thead>
-            <tbody>
-              {cars.map((car) => (
-                <tr key={car._id + "sodsfs"}>
-                  <td>{car.title}</td>
-                  <td className="amount">{car.price}</td>
-                  <td>
-                    <Link to="/">View</Link>
-                  </td>
-                  <td>
-                    <button
-                      onClick={() => handleOrderDelete(car._id)}
-                      className="delete_btn"
-                    >
-                      Delete
-                    </button>
-                  </td>
+          {cars.length === 0 ? (
+            <div
+              style={{ height: "60vh" }}
+              className="d-flex flex-column align-items-center justify-content-center"
+            >
+              <h4 style={{ color: "var(--main-color)" }}>Order Not found </h4>
+              <Link to="/car-listing" className="btn_round my-3 fs-6">
+                Go to Order
+              </Link>
+            </div>
+          ) : (
+            <table>
+              <thead>
+                <tr>
+                  <th>Car</th>
+                  <th>Amount</th>
+                  <th>Details</th>
+                  <th>Status</th>
                 </tr>
-              ))}
-            </tbody>
-          </table>
+              </thead>
+              <tbody>
+                {cars &&
+                  cars.map((car) => (
+                    <tr key={car._id + "sodsfs"}>
+                      <td>{car.title}</td>
+                      <td className="amount">{car.price}</td>
+                      <td>
+                        <Link to={`product-details/${car._id}`}>View</Link>
+                      </td>
+                      <td>
+                        <button
+                          onClick={() => handleOrderDelete(car._id)}
+                          className="delete_btn"
+                        >
+                          Delete
+                        </button>
+                      </td>
+                    </tr>
+                  ))}
+              </tbody>
+            </table>
+          )}
         </div>
       )}
     </Container>
